rename coreio to coreunix
