+++ core/coreunix/cat.go
@@ -1,15 +1,4 @@
-package core_io
-
-// TODO rename package to something that doesn't conflict with io/ioutil.
-// Pretty names are hard to find.
-//
-// Candidates:
-//
-// go-ipfs/core/unix
-// go-ipfs/core/io
-// go-ipfs/core/ioutil
-// go-ipfs/core/coreio
-// go-ipfs/core/coreunix
+package coreunix
 
 import (
 	"io"
